feat: fix up VRS mapping phase
